# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Data API Gateway
  description: "API Gateway multi-tenant (Maxtop, Pandora, Toplink) dengan dukungan Dual-DB (Postgres & MSSQL)."
  version: 1.0.0
...
servers:
  - url: /
    description: Current Server
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      description: API Key Developer
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token yang didapatkan dari login
  parameters:
    TenantParam:
      name: tenant
      in: path
      required: true
      description: Nama Unit/Tenant (maxtop, pandora, toplink)
      schema:
        type: string
    DbSourceHeader:
      name: X-DB-Source
      in: header
      required: false
      description: Pilih mesin database (default menggunakan postgres)
      schema:
        type: string
        enum: [postgres, mssql]
        default: postgres
  schemas:
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: "admin"
        password:
          type: string
          example: "secret123"
    CreateUserRequest:
      type: object
      required:
        - username
        - password
        - rules
      properties:
        username:
          type: string
          example: "op_user1"
        password:
          type: string
          example: "secret123"
        rules:
          type: string
          example: "op"
    UpdateUserRequest:
      type: object
      properties:
        password:
          type: string
          nullable: true
          example: "newsecret123"
        rules:
          type: string
          nullable: true
          example: "opout"
    Inbox:
      type: object
      properties:
        kode:
          type: integer
          format: int64
          example: 105423
        tgl_entri:
          type: string
          format: date-time
          example: "2026-09-06T10:00:00Z"
        tgl_status:
          type: string
          format: date-time
          nullable: true
          example: "2026-09-06T10:05:00Z"
        pengirim:
          type: string
          example: "08123456789"
        tipe_pengirim:
          type: string
          example: "SMS"
        penerima:
          type: string
          nullable: true
          example: "08987654321"
        pesan:
          type: string
          example: "SALDO"
        status:
          type: integer
          example: 0
        kode_terminal:
          type: integer
          nullable: true
          example: 1
        kode_reseller:
          type: string
          nullable: true
          example: "R001"
        kode_transaksi:
          type: integer
          nullable: true
          example: 54321
        is_jawaban:
          type: integer
          example: 0
        service_center:
          type: string
          nullable: true
          example: "SC01"
        is_cs:
          type: integer
          nullable: true
          example: 0
        kode_jawaban_cs:
          type: integer
          format: int64
          nullable: true
          example: 1234
        hash:
          type: string
          nullable: true
          example: "a1b2c3d4"
    CreateInboxRequest:
      type: object
      required:
        - pesan
        - pengirim
      properties:
        pesan:
          type: string
        pengirim:
          type: string
        penerima:
          type: string
        tipe_pengirim:
          type: string
        kode_terminal:
          type: integer
        kode_reseller:
          type: string
        kode_transaksi:
          type: integer
        service_center:
          type: string
        status:
          type: integer
        is_jawaban:
          type: integer
        is_cs:
          type: integer
        kode_jawaban_cs:
          type: string
        hash:
          type: string
    UpdateInboxRequest:
      type: object
      properties:
        pesan:
          type: string
          nullable: true
        status:
          type: integer
          nullable: true
        pengirim:
          type: string
          nullable: true
        tipe_pengirim:
          type: string
          nullable: true
        penerima:
          type: string
          nullable: true
        kode_terminal:
          type: integer
          nullable: true
        kode_reseller:
          type: string
          nullable: true
        kode_transaksi:
          type: integer
          nullable: true
        is_jawaban:
          type: integer
          nullable: true
        service_center:
          type: string
          nullable: true
        is_cs:
          type: integer
          nullable: true
        kode_jawaban_cs:
          type: integer
          format: int64
          nullable: true
        hash:
          type: string
          nullable: true
...
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          example: sukses
        data:
          type: object
          additionalProperties: true
...
    CursorPaginationMeta:
      type: object
      properties:
        has_next_page:
          type: boolean
        has_prev_page:
          type: boolean
        next_cursor:
          type: integer
          format: int64
    InboxResponse:
      type: object
      properties:
        trace_id:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/Inbox"
        meta:
          $ref: "#/components/schemas/CursorPaginationMeta"
...
paths:
...
  /api/v1/{tenant}/auth/login:
    post:
      summary: User Login
      description: Login dengan username dan password untuk mendapatkan token JWT.
      security:
        - ApiKeyAuth: []
      tags:
        - auth
      parameters:
        - $ref: "#/components/parameters/TenantParam"
        - $ref: "#/components/parameters/DbSourceHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        <<: *errorResponses
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /api/v1/{tenant}/auth/users:
    post:
      summary: Create User
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      tags:
        - user
      parameters:
        - $ref: "#/components/parameters/TenantParam"
        - $ref: "#/components/parameters/DbSourceHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        <<: *errorResponses
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /api/v1/{tenant}/auth/users/{username}:
    put:
      summary: Update User
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      tags:
        - user
      parameters:
        - $ref: "#/components/parameters/TenantParam"
        - $ref: "#/components/parameters/DbSourceHeader"
        - name: username
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        <<: *errorResponses
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /api/v1/{tenant}/inbox:
    get:
      summary: Get Inbox (13 Opsi Filter)
      description: "Mengambil daftar pesan Inbox"
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      tags:
        - inbox
      parameters:
        - $ref: "#/components/parameters/TenantParam"
        - $ref: "#/components/parameters/DbSourceHeader"
        - name: startDate
          in: query
          description: Tanggal mulai (Format YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: Tanggal akhir (Format YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: pageSize
          in: query
          description: Jumlah data per halaman (default 10, maksimal 500)
          schema:
            type: integer
        - name: limit
          in: query
          description: Total data maksimal yang boleh di-deliver selama pagination (opsional, maksimum 100.000)
          schema:
            type: integer
        - name: terminal
          in: query
          description: Filter kode terminal
          schema:
            type: integer
        - name: reseller
          in: query
          description: Filter kode reseller (kode persis / dropdown)
          schema:
            type: string
        - name: pengirim
          in: query
          description: Filter pengirim (pencocokan sebagian / ILIKE)
          schema:
            type: string
        - name: tipe
          in: query
          description: Filter tipe pengirim
          schema:
            type: string
        - name: status
          in: query
          description: Filter status pesan
          schema:
            type: integer
        - name: pesan
          in: query
          description: Pencarian kata kunci pada kolom pesan (LIKE)
          schema:
            type: string
        - name: requestFromReseller
          in: query
          description: Checkbox filter request dari reseller (kode_reseller IS NOT NULL)
          schema:
            type: boolean
        - name: jawabanFromProvider
          in: query
          description: Checkbox filter jawaban dari provider (is_jawaban = 1)
          schema:
            type: boolean
        - name: cursor
          in: query
          description: Cursor pagination (kode record terakhir)
          schema:
            type: integer
            format: int64
      responses:
        <<: *errorResponses
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboxResponse"
    post:
      summary: Insert Inbox
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      tags:
        - inbox
      parameters:
        - $ref: "#/components/parameters/TenantParam"
        - $ref: "#/components/parameters/DbSourceHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInboxRequest"
      responses:
        <<: *errorResponses
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /api/v1/{tenant}/inbox/{kode}:
    put:
      summary: Update Inbox
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      tags:
        - inbox
      parameters:
        - $ref: "#/components/parameters/TenantParam"
        - $ref: "#/components/parameters/DbSourceHeader"
        - name: kode
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInboxRequest"
      responses:
        <<: *errorResponses
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...